moved all the fragments out of the Main Activity and into seperate classes
