+++ ProjectSpartanV2/app/build.gradle
