+++ ProjectSpartanV2/app/src/main/java/com/haloproject/projectspartanv2/view/TopBar.java
@@ -1,4 +1,4 @@
-package com.haloproject.projectspartanv2;
+package com.haloproject.projectspartanv2.view;
 
 import android.content.Context;
 import android.content.res.TypedArray;
@@ -9,6 +9,8 @@ import android.graphics.drawable.Drawable;
 import android.text.TextPaint;
 import android.util.AttributeSet;
 import android.view.View;
+
+import com.haloproject.projectspartanv2.R;
 
 
 /**
