+++ ProjectSpartanV2/app/src/main/res/layout/fragment_battery.xml
@@ -13,7 +13,7 @@
             app:batteryCharge="0"
             android:layout_width="480px"
             android:layout_height="700px"
-            class="com.haloproject.projectspartanv2.BatteryBar"
+            class="com.haloproject.projectspartanv2.view.BatteryBar"
             android:id="@+id/highampbattery"
             android:layout_gravity="center_vertical" />
 
@@ -34,7 +34,7 @@
         android:layout_gravity="center_vertical">
 
         <view
-            class="com.haloproject.projectspartanv2.BatteryBar"
+            class="com.haloproject.projectspartanv2.view.BatteryBar"
             app:batteryCharge="0"
             android:layout_width="480px"
             android:layout_height="700px"
@@ -57,7 +57,7 @@
         android:layout_gravity="center_vertical">
 
         <view
-            class="com.haloproject.projectspartanv2.BatteryBar"
+            class="com.haloproject.projectspartanv2.view.BatteryBar"
             app:batteryCharge="80"
             android:layout_width="480px"
             android:layout_height="700px"
@@ -80,7 +80,7 @@
         android:layout_gravity="center_vertical">
 
         <view
-            class="com.haloproject.projectspartanv2.BatteryBar"
+            class="com.haloproject.projectspartanv2.view.BatteryBar"
             app:batteryCharge="0"
             android:layout_width="480px"
             android:layout_height="700px"
