+++ projectspartan/app/src/main/gen/org/unsc/projectspartan/Manifest.java
@@ -1,0 +1,7 @@
+/*___Generated_by_IDEA___*/
+
+package org.unsc.projectspartan;
+
+/* This stub is only used by the IDE. It is NOT the Manifest class actually packed into the APK */
+public final class Manifest {
+}
