+++ projectspartan/app/src/main/gen/org/unsc/projectspartan/R.java
@@ -1,0 +1,7 @@
+/*___Generated_by_IDEA___*/
+
+package org.unsc.projectspartan;
+
+/* This stub is only used by the IDE. It is NOT the R class actually packed into the APK */
+public final class R {
+}
